Rearrange `certifier` module
- `interface`: contains the trait & error definition
- `certifier_service`: the implementation
